move components to the component folder
