--- app/ui/button.module.css
@@ -1,26 +1,0 @@
-.radixPrimary{
-    /*base*/
-    --accent-9: rgba(var(--purple-rgb));
-    --cursor-button: pointer;
-    /*on hover /active */
-    --accent-10: var(--purple-hover);
-
-    /*on disable*/
-    --gray-a3: rgba(var(--purple-rgb), 0.25);
-    --gray-a8: var(--white);
-}
-.radixSecondary{
-    /*base*/
-    --accent-a11: var(--purple);
-    --accent-a8: var(--purple);
-    --cursor-button: pointer;
-    /*on hover /active */
-    --accent-a2:   var(--light-purple);
-    --accent-a3:   var(--light-purple);
-    /*on disable*/
-    --gray-a8: rgba(var(--purple-rgb), 0.25);
-    --gray-a7: rgba(var(--purple-rgb), 0.25);
-}
-
-/*
-*/
